requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.1
Werkzeug==3.0.1
python-dotenv==1.0.0
Pillow>=10.0.0
email-validator==2.1.0
markdown==3.5.1
bleach==6.1.0
qrcode>=7.4.2
python-barcode>=0.15.1
weasyprint>=60.0
reportlab>=4.0.0